+++ Wrappers/Python/cil/optimisation/operators/FiniteDifferenceOperator.py
@@ -19,6 +19,7 @@
 import numpy as np
 
 from cil.optimisation.operators import LinearOperator
+from cil.utilities.errors import InPlaceError
         
 class FiniteDifferenceOperator(LinearOperator):
     
@@ -84,6 +85,9 @@ class FiniteDifferenceOperator(LinearOperator):
         return tmp       
 
     def direct(self, x, out = None):
+        
+        if id(x)==id(out):
+            raise InPlaceError(message="FiniteDifferenceOperator.direct cannot be used in place")
         
         x_asarr = x.as_array()
         
@@ -224,6 +228,9 @@ class FiniteDifferenceOperator(LinearOperator):
                  
         
     def adjoint(self, x, out=None):
+        
+        if id(x)==id(out):
+            raise InPlaceError
         
         # Adjoint operation defined as