+++ Wrappers/Python/test/test_Operator.py
@@ -33,6 +33,7 @@ from cil.optimisation.operators import SumOperator,  ZeroOperator, CompositionOp
 from cil.utilities import dataexample
 import logging
 from testclass import CCPiTestClass
+from cil.utilities.errors import InPlaceError
 
 
 initialise_tests()
@@ -42,9 +43,7 @@ def dt(steps):
 
 
 class TestOperator(CCPiTestClass):
-    def setUp(self):
-        numpy.random.seed(1)
-
+  
 
     def test_MatrixOperator(self):        
         m = 30